feature - update with env & key vault provider
