+++ src/Arcus.Security.Startup.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30011.22
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Arcus.Security.Startup", "Arcus.Security.Startup.csproj", "{8D332AEA-B722-46C7-A3E8-CD4D38E868D0}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{8D332AEA-B722-46C7-A3E8-CD4D38E868D0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8D332AEA-B722-46C7-A3E8-CD4D38E868D0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8D332AEA-B722-46C7-A3E8-CD4D38E868D0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8D332AEA-B722-46C7-A3E8-CD4D38E868D0}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {BFC7C120-EAA8-4952-AB96-0053B314D0EB}
+	EndGlobalSection
+EndGlobal
