+++ src/Security/EnvironmentVariableSecretProvider.cs
@@ -1,0 +1,32 @@
+﻿using System;
+using System.Threading.Tasks;
+using Arcus.Security.Core;
+
+namespace Arcus.Security.Startup.Security
+{
+    public class EnvironmentVariableSecretProvider : ISecretProvider
+    {
+        /// <summary>Retrieves the secret value, based on the given name</summary>
+        /// <param name="secretName">The name of the secret key</param>
+        /// <returns>Returns the secret key.</returns>
+        /// <exception cref="T:System.ArgumentException">The <paramref name="secretName" /> must not be empty</exception>
+        /// <exception cref="T:System.ArgumentNullException">The <paramref name="secretName" /> must not be null</exception>
+        /// <exception cref="T:Arcus.Security.Core.SecretNotFoundException">The secret was not found, using the given name</exception>
+        public Task<string> GetRawSecretAsync(string secretName)
+        {
+            return Task.FromResult(Environment.GetEnvironmentVariable(secretName));
+        }
+
+        /// <summary>Retrieves the secret value, based on the given name</summary>
+        /// <param name="secretName">The name of the secret key</param>
+        /// <returns>Returns a <see cref="T:Arcus.Security.Core.Secret" /> that contains the secret key</returns>
+        /// <exception cref="T:System.ArgumentException">The <paramref name="secretName" /> must not be empty</exception>
+        /// <exception cref="T:System.ArgumentNullException">The <paramref name="secretName" /> must not be null</exception>
+        /// <exception cref="T:Arcus.Security.Core.SecretNotFoundException">The secret was not found, using the given name</exception>
+        public async Task<Secret> GetSecretAsync(string secretName)
+        {
+            string secretValue = await GetRawSecretAsync(secretName);
+            return new Secret(secretValue, "1.0.0");
+        }
+    }
+}
