+++ src/Security/InMemorySecretProvider.cs
@@ -1,0 +1,48 @@
+﻿using System.Collections.Generic;
+using System.Threading.Tasks;
+using Arcus.Security.Core;
+using GuardNet;
+
+namespace Arcus.Security.Startup.Security
+{
+    public class InMemorySecretProvider : ISecretProvider
+    {
+        private readonly IDictionary<string, Secret> _secrets;
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="InMemorySecretProvider"/> class.
+        /// </summary>
+        public InMemorySecretProvider(IDictionary<string, Secret> secrets)
+        {
+            Guard.NotNull(secrets, nameof(secrets));
+            _secrets = secrets;
+        }
+
+        /// <summary>
+        /// Retrieves the secret value, based on the given name
+        /// </summary>
+        /// <param name="secretName">The name of the secret key</param>
+        /// <returns>Returns the secret key.</returns>
+        /// <exception cref="System.ArgumentException">The <paramref name="secretName"/> must not be empty</exception>
+        /// <exception cref="System.ArgumentNullException">The <paramref name="secretName"/> must not be null</exception>
+        /// <exception cref="SecretNotFoundException">The secret was not found, using the given name</exception>
+        public async Task<string> GetRawSecretAsync(string secretName)
+        {
+            Secret secret = await GetSecretAsync(secretName);
+            return secret?.Value;
+        }
+
+        /// <summary>
+        /// Retrieves the secret value, based on the given name
+        /// </summary>
+        /// <param name="secretName">The name of the secret key</param>
+        /// <returns>Returns a <see cref="Secret"/> that contains the secret key</returns>
+        /// <exception cref="System.ArgumentException">The <paramref name="secretName"/> must not be empty</exception>
+        /// <exception cref="System.ArgumentNullException">The <paramref name="secretName"/> must not be null</exception>
+        /// <exception cref="SecretNotFoundException">The secret was not found, using the given name</exception>
+        public Task<Secret> GetSecretAsync(string secretName)
+        {
+            return Task.FromResult(_secrets[secretName]);
+        }
+    }
+}
